+++ src/main/java/com/michaeladrummonds/aguafina/services/impl/ProductServiceImpl.java
@@ -35,7 +35,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Product getProductById(Long id) {
+    public Product getProductById(Integer id) {
         return productRepository.findById(id).get();
     }
